requirements.txt
Flask==2.3.2
pytesseract==0.3.10
Pillow==10.0.0
googletrans==4.0.0rc1
Flask-Cors


